backend/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
mysqlclient==2.2.6
sqlparse==0.5.3
